Correct transform origin for block fragmentation.
The spec [1] says that the transform origin is to be calculated
independently for each fragment. We'll now do this when
LayoutNGBlockFragmentation is enabled, unless fragment stitching is
required (i.e. we're compositing and want block fragmentation at the
same time).

Added two WPT tests - one that triggers compositing, and one that
doesn't. The one that triggers compositing will fail both with legacy
and NG block fragmentation, as long as CompositeAfterPaint isn't
enabled.

Since we need to stitch fragments together when CompositeAfterPaint
isn't enabled, I copied the compositing-triggering WPT test over to
fast/multicol/, so that we can keep an expectation file for the current
incorrect rendering, to ensure that we stay bug-compatible with legacy
block fragmentation.

[1] https://www.w3.org/TR/css-break-3/#transforms

Bug: 829028
Change-Id: I01eedaef80787d66926a2347131a7b0967f85ecc